+++ controller/ServicesController.js
@@ -27,7 +27,7 @@ class ServicesController {
     static async addFunctionality(req, res){
         try {
             if(!RolesController.isAdmin(req.session.userrole)){
-                res.redirect('/administracao');
+                res.redirect('/autenticacao');
                 return;
             }
             res.render(Utils.Titles.addFunctionality.name_page, { 'title': Utils.Titles.addFunctionality.title, 'allRoles': await RolesController.allRoles()});
@@ -39,15 +39,15 @@ class ServicesController {
     static async addFunctionalityPost(req, res){
         try {
             if(!RolesController.isAdmin(req.session.userrole)){
-                res.redirect('/administracao');
+                res.redirect('/autenticacao');
                 return;
             }
             const { name, value, RoleId } = req.body;
             await __newService(name, value, RoleId);
-            Utils.Message.redirect(req, res, '/administracao/adicionar_servico', 'Serviço adicionado com sucesso.');
+            Utils.Message.redirect(req, res, '/funcionalidade/adicionar_servico', 'Serviço adicionado com sucesso.');
         } catch (error) {
             if(error.errors[0].type && error.errors[0].type === 'unique violation'){
-                Utils.Message.redirect(req, res, '/administracao/adicionar_servico', 'Serviço já existente, tente outro nome para o serviço.');
+                Utils.Message.redirect(req, res, '/funcionalidade/adicionar_servico', 'Serviço já existente, tente outro nome para o serviço.');
             }
         }
     }
@@ -55,7 +55,7 @@ class ServicesController {
     static async removeFunctionality(req, res){
         try {
             if(!RolesController.isAdmin(req.session.userrole)){
-                res.redirect('/administracao');
+                res.redirect('/autenticacao');
                 return;
             }
             res.render(Utils.Titles.removeFunctionality.name_page, { 'title': Utils.Titles.removeFunctionality.title});
@@ -67,15 +67,30 @@ class ServicesController {
     static async removeFunctionalityPost(req, res){
         try {
             if(!RolesController.isAdmin(req.session.userrole)){
-                res.redirect('/administracao');
+                res.redirect('/autenticacao');
                 return;
             }
             const { name, value, RoleId } = req.body;
             await __removeService(name);
-            Utils.Message.redirect(req, res, '/administracao/remover_servico', 'Serviço removido com sucesso.');
+            Utils.Message.redirect(req, res, '/funcionalidade/remover_servico', 'Serviço removido com sucesso.');
         } catch (error) {
             console.log(error)
         }
+    }
+
+    //a fazer
+    static async functionalities(req, res) {
+        res.render(Utils.Titles.functionalities.name_page, { 'title': Utils.Titles.functionalities.title });
+    }
+
+    //a fazer
+    static async panel(req, res) {
+        res.render(Utils.Titles.panel.name_page, { 'title': Utils.Titles.panel.title });
+    }
+
+    //a fazer
+    static async panelPost(req, res) {
+        res.render(Utils.Titles.panel.name_page, { 'title': Utils.Titles.panel.title });
     }
 };
