+++ routes/AllRoutes.js
@@ -1,10 +1,13 @@
 const express = require('express');
 const router = express.Router();
 const HomeController = require('../controller/HomeController');
-const admin_routes = require('./admin');
+const authentication_routes = require('./authentication');
+const functionalities_routes = require('./functionalities');
 
-router.use('/administracao', admin_routes);
+router.get('/', HomeController.render);
 
-router.get('/', HomeController.render)
+router.use('/autenticacao', authentication_routes);
+
+router.use('/funcionalidade', functionalities_routes);
 
 module.exports = router;
