+++ views/addFunctionality.handlebars
@@ -3,7 +3,7 @@
         <div class="row justify-content-start">
             <h2>Adicionar serviço</h2>
             <p class="text-secondary">Você pode adicionar serviços a tipo de funcionario.</p>
-            <form action="/administracao/adicionar_servico" method="post" class="text-start col-md-6">
+            <form action="/funcionalidade/adicionar_servico" method="post" class="text-start col-md-6">
                 <label for="name" class="form-label m-1">Nome</label>
                 <input type="text" id="name" name="name" class="form-control m-1" required>
                 <label for="value" class="form-label m-1">Preço</label>
