+++ views/addUser.handlebars
@@ -3,7 +3,7 @@
         <div class="row justify-content-start">
             <h2>Adicionar funcionario</h2>
             <p class="text-secondary">Você pode adicionar os funcionarios do seu petshop.</p>
-            <form action="/administracao/adicionar_funcionario" method="post" class="text-start col-md-6">
+            <form action="/funcionalidade/adicionar_funcionario" method="post" class="text-start col-md-6">
                 <label for="full_name" class="form-label m-1">Nome completo</label>
                 <input type="text" id="full_name" name="full_name" value="{{full_name}}" class="form-control m-1"
                     required>
