+++ views/login.handlebars
@@ -5,13 +5,13 @@
             <p class="text-secondary">A área administrativa é o seu hub central para gerenciar o PetShop Amigo
                 Fiel.</br>Faça
                 login para acessar ferramentas e recursos exclusivos.</p>
-            <form action="/administracao" method="post" class="text-start col-md-6">
+            <form action="/autenticacao" method="post" class="text-start col-md-6">
                 <label for="email" class="form-label m-1">E-mail</label>
                 <input type="email" id="email" name="email" class="form-control m-1" required>
                 <label for="password" class="form-label m-1">Senha</label>
                 <input type="password" id="password" name="password" class="form-control m-1" required>
                 <div class="d-flex justify-content-between my-1">
-                    <a href="/administracao/recuperar_senha">Esqueceu a senha?</a>
+                    <a href="/autenticacao/recuperar_senha">Esqueceu a senha?</a>
                 </div>
                 <button type="submit" class="btn btn-primary my-2 w-100">Entrar em sua conta</button>
             </form>
