+++ views/partials/footer.handlebars
@@ -12,7 +12,7 @@
             <a href="/#Servicos" class="text-decoration-none text-light">Serviços</a>
             <a href="/#Sobre" class="text-decoration-none text-light">Sobre</a>
             <a href="/#Contato" class="text-decoration-none text-light">Contate-nos</a>
-            <a href="/administracao" class="text-decoration-none text-light">Administração</a>
+            <a href="/autenticacao" class="text-decoration-none text-light">Administração</a>
         </nav>
     </div>
     <div class="col-md-3 my-2">
