+++ views/partials/navbar_logged.handlebars
@@ -1,7 +1,7 @@
 <header>
         <nav class="navbar navbar-expand-lg bg-body-tertiary">
             <div class="container-fluid">
-                <a class="navbar-brand" href="/administracao/painel">[Nome da empresa]</a>
+                <a class="navbar-brand" href="/funcionalidade/painel">[Nome da empresa]</a>
                 <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                     data-bs-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false"
                     aria-label="Toggle navigation">
@@ -9,9 +9,9 @@
                 </button>
                 <div class="collapse navbar-collapse justify-content-end" id="navbarNavAltMarkup">
                     <div class="navbar-nav">
-                        <a class="nav-link" aria-current="page" href="/administracao/painel">Painel</a>
-                        <a class="nav-link" aria-current="page" href="/administracao/funcionalidades">Funcionalidades</a>
-                        <a class="nav-link" aria-current="page" href="/administracao/usuario">Usuário</a>
+                        <a class="nav-link" aria-current="page" href="/funcionalidade/painel">Painel</a>
+                        <a class="nav-link" aria-current="page" href="/funcionalidade/listar_funcionalidades">Funcionalidades</a>
+                        <a class="nav-link" aria-current="page" href="/funcionalidade/ver_perfil">Usuário</a>
                     </div>
                 </div>
             </div>
