+++ views/removeFunctionality.handlebars
@@ -3,7 +3,7 @@
         <div class="row justify-content-start">
             <h2>Remover serviço</h2>
             <p class="text-secondary">Você pode remover serviços.</p>
-            <form action="/administracao/remover_servico" method="post" class="text-start col-md-6">
+            <form action="/funcionalidade/remover_servico" method="post" class="text-start col-md-6">
                 <label for="name" class="form-label m-1">Nome</label>
                 <input type="text" id="name" name="name" class="form-control m-1" required>
                 <div class="d-flex justify-content-between my-1">
