+++ views/removeUser.handlebars
@@ -3,7 +3,7 @@
         <div class="row justify-content-start">
             <h2>Remover funcionario</h2>
             <p class="text-secondary">Remover os funcionarios do seu petshop.</p>
-            <form action="/administracao/remover_funcionario" method="post" class="text-start col-md-6">
+            <form action="/funcionalidade/remover_funcionario" method="post" class="text-start col-md-6">
                 <label for="select_employee" class="form-label m-1">Selecione qual funcionario vai remover</label>
                 <select id="select_employee" name="select_employee" class="form-select form-select-lg mb-3" aria-label="Selecione qual funcionario vai remover" required>
                     {{#each allUsers}}
