+++ views/usuario.handlebars
@@ -3,7 +3,7 @@
         <div class="row justify-content-start">
             <h2>Usuário</h2>
             <p class="text-secondary">Você pode alterar seus dados cadastrados no sistema.</p>
-            <form action="/administracao/atualizar_usuario" method="post" class="text-start col-md-6">
+            <form action="/funcionalidade/atualizar_perfil" method="post" class="text-start col-md-6">
                 <label for="full_name" class="form-label m-1">Nome completo</label>
                 <input type="text" id="full_name" name="full_name" value="{{full_name}}" class="form-control m-1">
                 <label for="cpf" class="form-label m-1">CPF</label>
@@ -14,7 +14,7 @@
                 <input type="password" id="password" name="password" class="form-control m-1">
                 <div class="d-flex justify-content-between my-1">
                     <button type="submit" class="btn btn-primary my-2">Salvar alterações</button>    
-                    <a class="btn btn-warning my-2" href="/administracao/remover_autenticacao">Sair da conta</a>
+                    <a class="btn btn-warning my-2" href="/autenticacao/remover">Sair da conta</a>
                 </div>
             </form>
         </div>
